requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
black==23.10.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
defusedxml==0.7.1
ecdsa==0.18.0
factory-boy==3.3.0
Faker==24.3.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.4
httpie==3.2.2
httptools==0.6.1
httpx==0.27.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.3.2
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
passlib==1.7.4
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
pygraphviz==1.12
PyMySQL==1.1.0
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.4.3
pytest-asyncio==0.23.6
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.28
sqlalchemy2-stubs==0.0.2a36
sqlmodel==0.0.11
starlette==0.27.0
threadpoolctl==3.3.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.1.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.37.1
